requirements.txt
aiofiles==23.1.0
aiohttp==3.8.3
aiosignal==1.3.1
altair==4.2.2
anyio==3.6.2
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
Automat==22.10.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==2.1.1
chromedriver-autoinstaller==0.4.0
click==8.1.3
constantly==15.1.0
contourpy==1.0.7
cryptography==39.0.1
cssselect==1.2.0
cycler==0.11.0
discord==2.2.2
discord.py==2.2.2
entrypoints==0.4
et-xmlfile==1.1.0
exceptiongroup==1.1.0
fastapi==0.92.0
ffmpy==0.3.0
filelock==3.9.0
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2023.1.0
gradio==3.19.1
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
itemadapter==0.7.0
itemloaders==1.0.6
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsonschema==4.17.3
kiwisolver==1.4.4
linkify-it-py==2.0.0
lxml==4.9.2
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.6.3
mdit-py-plugins==0.3.3
mdurl==0.1.2
multidict==6.0.4
numpy==1.24.2
openai==0.27.2
openpyxl==3.1.0
orjson==3.8.6
outcome==1.2.0
packaging==23.0
pandas==1.5.3
parsel==1.7.0
Pillow==9.4.0
plotly==5.13.0
Protego==0.2.1
psycopg2-binary==2.9.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pycryptodome==3.17
pydantic==1.10.5
PyDispatcher==2.0.6
pydub==0.25.1
pyOpenSSL==23.0.0
pyparsing==3.0.9
pyrsistent==0.19.3
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==0.21.1
python-multipart==0.0.5
pytz==2022.7.1
PyYAML==6.0
queuelib==1.6.2
requests==2.28.2
requests-file==1.5.1
rfc3986==1.5.0
scikit-learn==1.2.1
scipy==1.10.0
Scrapy==2.8.0
selenium==4.8.0
service-identity==21.1.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.3
starlette==0.25.0
tenacity==8.2.1
threadpoolctl==3.1.0
tldextract==3.4.0
toolz==0.12.0
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
Twisted==22.10.0
typing_extensions==4.4.0
uc-micro-py==1.0.1
urllib3==1.26.14
uvicorn==0.20.0
w3lib==2.1.1
websockets==10.4
wsproto==1.2.0
yarl==1.8.2
zope.interface==5.5.2
